+++ onegov/election_day/tests/views/test_views_upload_election.py
@@ -471,7 +471,7 @@ def test_upload_election_available_formats_municipality(election_day_app_bern):
     assert [o[0] for o in upload.form['file_format'].options] == ['internal']
 
 
-def test_upload_election_notify_hipchat(election_day_app):
+def test_upload_election_notify_zulip(election_day_app):
 
     client = Client(election_day_app)
     client.get('/locale/de_CH').follow()
@@ -528,7 +528,7 @@ def test_upload_election_notify_hipchat(election_day_app):
 
     with patch('urllib.request.urlopen') as urlopen:
 
-        # Hipchat not set
+        # No settings
         upload = client.get('/election/election/upload').follow()
         upload.form['file_format'] = 'internal'
         upload.form['results'] = Upload('data.csv', csv, 'text/plain')
@@ -538,8 +538,10 @@ def test_upload_election_notify_hipchat(election_day_app):
 
         assert not urlopen.called
 
-        election_day_app.hipchat_token = 'abcd'
-        election_day_app.hipchat_room_id = '1234'
+        election_day_app.zulip_url = 'https://xx.zulipchat.com/api/v1/messages'
+        election_day_app.zulip_stream = 'WAB'
+        election_day_app.zulip_user = 'wab-bot@seantis.zulipchat.com'
+        election_day_app.zulip_key = 'aabbcc'
 
         upload = client.get('/election/election/upload').follow()
         upload.form['file_format'] = 'internal'
@@ -549,7 +551,7 @@ def test_upload_election_notify_hipchat(election_day_app):
         sleep(5)
 
         assert urlopen.called
-        assert 'api.hipchat.com' in urlopen.call_args[0][0].get_full_url()
+        assert 'xx.zulipchat.com' in urlopen.call_args[0][0].get_full_url()
 
 
 def test_upload_election_submit(election_day_app):